\name{rqapp-package}
\alias{rqapp-package}
\alias{rqapp}
\docType{package}
\title{
  Recurrence Quantification Analysis
}
\description{
  This package peforms several variants of recurrence quantification analysis
  (RQA) on both categorical and continuous time series. In addition, the 
  package provides the capability of performing cross RQA, joint RQA, and
  multidimensional RQA. In each case, the package calculates the standard 
  battery of recurrence plot metrics (e.g., recurrence rate, determinism,
  trapping time, etc...). The package also follows Coco and Dale (2015)
  by allowing the user to profile cross-recurrence plots by only looking at 
  diagonal recurrence points.  
}
\author{
Aaron D. Likens & Laura K. Allen

Maintainer: Aaron Likens <aaron.d.likens@gmail.com>
}
\references{
	Marwan, N., Romano, M. C. Theil, M., & Kurths, J. (2007). Recurrence plots
	    for the analysis of complex systems. \emph{Physics Reports}, \emph{438},
	    237-329.
	    
  	Wallot, S., Roepstorff, A., & Mønster, D. (2016). Multidimensional 
    	Recurrence Quantification Analysis (MdRQA) for the analysis of 
	    multidimensional time-series: A software implementation in MATLAB and
	    its	application to group-level data in joint action. \emph{Frontiers in
	    psychology}, \emph{7}, 1835.
	    
	Webber, C. L., & Zbilut, J. P. (2005). Recurrence quantification analysis
		of nonlinear dynamical time series. In S. Riley and G. C. Van Orden
		(eds). \emph{Tutorials in contemporary nonlinear methods for the
		behavioral sciences}.
	

}
\keyword{ package }
\seealso{
  rqa
}

